requirements.txt
langchain
langchain-openai
langchain-community
openai
jira
python-dotenv
pandas
faiss-cpu
matplotlib
scikit-learn